# py-genai/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Hello-GenAI API",
    "description": "REST API for the Hello-GenAI local chat application powered by Docker Model Runner.",
    "version": "2.0.0"
  },
  "servers": [
    { "url": "/", "description": "Current server" }
  ],
...
  "components": {
    "parameters": {
      "session_id": {
        "name": "session_id",
        "in": "path",
        "required": true,
        "description": "UUID of the session",
        "schema": { "type": "string", "format": "uuid" },
        "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "Message is required" }
        }
      },
      "Ok": {
        "type": "object",
        "properties": {
          "ok": { "type": "boolean", "example": true }
        }
      },
...
      "Preset": {
        "type": "object",
        "example": { "id": 1, "name": "Concise", "text": "You are a concise assistant.", "created_at": "2024-01-01T00:00:00+00:00" },
        "properties": {
          "id":         { "type": "integer" },
          "name":       { "type": "string", "maxLength": 60 },
          "text":       { "type": "string", "maxLength": 2000 },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Memory": {
        "type": "object",
        "example": {
          "id": 1,
          "content": "User prefers concise answers with code examples",
          "source_session_id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
          "enabled": 1,
          "created_at": "2024-01-01T00:00:00+00:00",
          "updated_at": "2024-01-01T00:00:00+00:00"
        },
        "properties": {
          "id":                { "type": "integer" },
          "content":           { "type": "string", "maxLength": 300 },
          "source_session_id": { "type": "string", "format": "uuid", "nullable": true, "description": "Session the fact was extracted from; null for manually added memories" },
          "enabled":           { "type": "integer", "enum": [0, 1], "description": "Disabled memories stay stored but are not recalled" },
          "created_at":        { "type": "string", "format": "date-time" },
          "updated_at":        { "type": "string", "format": "date-time" }
        }
      },
...
      "ChatRequest": {
        "type": "object",
        "required": ["message"],
        "example": {
          "message": "What is a container?",
          "session_id": null,
          "model": null,
          "system_prompt": null,
          "temperature": null,
          "max_tokens": null
        },
        "properties": {
          "message":       { "type": "string", "maxLength": 32000, "example": "What is a container?", "description": "Max length configurable via MAX_MESSAGE_LEN (default 32000)" },
          "session_id":    { "type": "string", "format": "uuid", "nullable": true, "description": "Omit to chat without session history" },
          "model":         { "type": "string", "nullable": true, "example": "docker.io/ai/gemma4:latest", "description": "Overrides the server default model" },
          "system_prompt": { "type": "string", "nullable": true, "description": "Overrides the session or default system prompt" },
          "temperature":   { "type": "number", "minimum": 0, "maximum": 2, "nullable": true, "example": 0.7, "description": "Sampling temperature, clamped server-side to 0–2. Omit to use the model default." },
          "max_tokens":    { "type": "integer", "minimum": 1, "maximum": 32768, "nullable": true, "example": 512, "description": "Maximum response length in tokens, clamped server-side to 1–32768. Omit to use the model default." },
          "save":          { "type": "boolean", "default": true, "description": "Set false to run with session context but persist nothing (used by compare mode)" },
          "images":        { "type": "array", "items": { "type": "string" }, "description": "Inline image data URLs (data:image/…) for vision-capable models", "example": [] },
          "use_memory":    { "type": "boolean", "default": true, "description": "Set false to skip recalling stored memories and extracting new facts for this request" },
          "use_tools":     { "type": "boolean", "default": true, "description": "Allow the model to call built-in tools (requires backend tool support)" },
          "use_rag":       { "type": "boolean", "default": true, "description": "Retrieve relevant knowledge-base passages into the prompt" },
          "regenerate":    { "type": "boolean", "default": false, "description": "Regenerate as a new branch under parent_message_id instead of a new turn" },
          "parent_message_id": { "type": "integer", "nullable": true, "description": "The user message id to branch the regenerated response under" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/presets": {
      "get": {
        "tags": ["Presets"],
        "summary": "List system prompt presets",
        "responses": {
          "200": {
            "description": "Presets ordered by name",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Preset" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Presets"],
        "summary": "Create a preset",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "text"],
                "properties": {
                  "name": { "type": "string", "maxLength": 60, "example": "Concise" },
                  "text": { "type": "string", "maxLength": 2000, "example": "You are a concise assistant." }
                }
              },
              "example": { "name": "Concise", "text": "You are a concise assistant." }
            }
          }
        },
        "responses": {
          "201": { "description": "Preset created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Preset" } } } },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/presets/{preset_id}": {
      "delete": {
        "tags": ["Presets"],
        "summary": "Delete a preset",
        "parameters": [
          {
            "name": "preset_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" },
            "example": 1
          }
        ],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } }
        }
      }
    },
...
    "/api/stream": {
      "post": {
        "tags": ["Chat"],
        "summary": "Streaming chat (SSE)",
        "description": "Streams the response as Server-Sent Events. Three event types are emitted:\n\n- `start` — fired once before tokens; includes `user_message_id`\n- `token` — one chunk of streamed text\n- `done` — stream complete; includes `usage`, `message_id`, `is_first`\n- `error` — error string",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatRequest" },
              "example": {
                "message": "Explain Docker networking in simple terms."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE stream",
            "content": {
              "text/event-stream": {
                "schema": { "type": "string" },
                "example": "data: {\"start\": true, \"user_message_id\": 1}\n\ndata: {\"token\": \"Hello\"}\n\ndata: {\"done\": true, \"usage\": {}, \"message_id\": 2, \"is_first\": true}\n\n"
              }
            }
          },
          "400": { "description": "Validation error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/api/models": {
      "get": {
        "tags": ["Models"],
        "summary": "List available models",
        "description": "Discovers the models the LLM backend currently exposes (live). `AVAILABLE_MODELS` / `LLAMA_MODEL` are used only as a fallback when the backend is unreachable. Responses are cached for 30s; pass any query string (e.g. `?refresh=<ts>`) to bypass the cache.",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "description": "Any value bypasses the 30s response cache and forces live re-discovery",
            "schema": { "type": "string" },
            "example": "1720000000000"
          }
        ],
        "responses": {
          "200": {
            "description": "Model list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": {
                    "models": ["docker.io/ai/gemma4:latest", "docker.io/ai/gemma3n:latest"],
                    "current": "docker.io/ai/gemma4:latest",
                    "source": "live"
                  },
                  "properties": {
                    "models":  { "type": "array", "items": { "type": "string" } },
                    "current": { "type": "string", "description": "Suggested default: the configured model if available, else the first model" },
                    "source":  { "type": "string", "enum": ["live", "fallback"], "description": "`live` = discovered from the backend; `fallback` = backend unreachable, using configured models" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/memories": {
      "get": {
        "tags": ["Memory"],
        "summary": "List memories",
        "description": "All stored facts about the user, including disabled ones.",
        "responses": {
          "200": {
            "description": "Memories in creation order",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Memory" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Memory"],
        "summary": "Add a memory manually",
        "description": "Duplicate content (case-insensitive) is not stored twice — the existing memory's id is returned.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": { "type": "string", "maxLength": 300, "example": "User's name is Anirban" }
                }
              },
              "example": { "content": "User's name is Anirban" }
            }
          }
        },
        "responses": {
          "201": { "description": "Memory stored", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Memory" } } } },
          "400": { "description": "Missing content", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "409": { "description": "Memory is full (MEMORY_MAX_ITEMS reached)", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["Memory"],
        "summary": "Forget everything",
        "description": "Deletes all stored memories.",
        "responses": {
          "200": {
            "description": "Cleared",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "example": { "ok": true, "deleted": 7 },
                  "properties": {
                    "ok":      { "type": "boolean" },
                    "deleted": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/memories/{memory_id}": {
      "parameters": [
        {
          "name": "memory_id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer" },
          "example": 1
        }
      ],
      "patch": {
        "tags": ["Memory"],
        "summary": "Edit or pause a memory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": { "type": "string", "maxLength": 300 },
                  "enabled": { "type": "boolean", "description": "false pauses the memory without deleting it" }
                }
              },
              "example": { "enabled": false }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } },
          "400": { "description": "Invalid content", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["Memory"],
        "summary": "Forget one memory",
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } }
        }
      }
    },
...
    "/api/projects": {
      "get": {
        "tags": ["Projects"], "summary": "List projects",
        "responses": { "200": { "description": "Projects with session counts" } }
      },
      "post": {
        "tags": ["Projects"], "summary": "Create a project",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["name"],
          "properties": { "name": { "type": "string" }, "system_prompt": { "type": "string", "nullable": true } }
        }, "example": { "name": "Work", "system_prompt": null } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Missing name" } }
      }
    },
    "/api/projects/{project_id}": {
      "parameters": [ { "name": "project_id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 } ],
      "patch": {
        "tags": ["Projects"], "summary": "Rename or set a project's system prompt",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "properties": { "name": { "type": "string" }, "system_prompt": { "type": "string", "nullable": true } }
        } } } },
        "responses": { "200": { "description": "Updated" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "tags": ["Projects"], "summary": "Delete a project",
        "description": "Its sessions become unfiled (project_id → null); scoped memories and documents are removed.",
        "responses": { "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } } }
      }
    },
    "/api/sessions/{session_id}/project": {
      "post": {
        "tags": ["Projects"], "summary": "Assign a session to a project",
        "parameters": [ { "$ref": "#/components/parameters/session_id" } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "properties": { "project_id": { "type": "integer", "nullable": true } }
        }, "example": { "project_id": 1 } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } } }
      }
    },
...
    "/api/templates": {
      "get": { "tags": ["Templates"], "summary": "List slash-command templates", "responses": { "200": { "description": "Templates" } } },
      "post": {
        "tags": ["Templates"], "summary": "Create a template",
        "requestBody": { "required": true, "content": { "application/json": { "schema": {
          "type": "object", "required": ["trigger", "content"],
          "properties": { "trigger": { "type": "string", "example": "tldr" }, "title": { "type": "string" }, "content": { "type": "string" } }
        } } } },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Missing fields" } }
      }
    },
    "/api/templates/{template_id}": {
      "delete": {
        "tags": ["Templates"], "summary": "Delete a template",
        "parameters": [ { "name": "template_id", "in": "path", "required": true, "schema": { "type": "integer" } } ],
        "responses": { "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } } }
      }
    },
...
  }
}